twitch popout frame done
needed 0 js, because border-box and giving the player a padding is a
great hack. it's doesn't fully prevent black bars yet.
I also added a little nowlive ad to the twitch page and did some styling
for that section (repeated in index.php and youtube.php) and removed all
leftover debugDIVs