Removed WatchLater playlist support, Stylized Radio
as /v/videoID is no longer supported by youtube, the feature got
removed, therefore simplyfying the form and code a lot. I also added a
bunch of spacing, while looking through the file.

The new stylized Radio inputs can be found on the twitchpage. They
require a small amount of css and js in the respective html file (which
in return adds the ability to change the width) and the wrapper div's id
has to be added in the 2 marked lines in the css file.

radioValue() now automatically uses names, so the input is now a string
with the shared name for the radio inputs in question.