re-did navigation, fixed the little footer bug
decided to try my inital plan again and figured it out.

I also fixed the little bug caused by the git icon. It was caused by the font-size being 14, so 2px smaller than the image. I fixed it by moving the image to the background of `::before`, putting that into absolute position mode and giving it space by adding a padding to the `<a>` element